+++ samples/Demo/Beef.Demo.Database/Migrations/20210427-143007-create-demo-eventoutbox.sql
@@ -1,0 +1,10 @@
+CREATE TABLE [Demo].[EventOutbox] (
+  /*
+   * This is automatically generated; any changes will be lost.
+   */
+
+  [EventOutboxId] BIGINT IDENTITY (1, 1) NOT NULL PRIMARY KEY NONCLUSTERED ([EventOutboxId] ASC),
+  [EnqueuedDate] DATETIME NOT NULL,
+  [DequeuedDate] DATETIME NULL,
+  CONSTRAINT [IX_Demo__DequeuedDate] UNIQUE CLUSTERED ([DequeuedDate], [EventOutboxId])
+);